Move field factory classes to separate namespace
