+++ tests/PureFieldTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\Form\Tests;
 
 use PHPUnit\Framework\TestCase;
-use Yiisoft\Form\PureField;
+use Yiisoft\Form\Field\Factory\PureField;
 use Yiisoft\Form\Tests\Support\ThemedPureField;
 use Yiisoft\Form\Theme\ThemeContainer;
 use Yiisoft\Html\Tag\Button;
