+++ tests/Support/ThemedPureField.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\Form\Tests\Support;
 
-use Yiisoft\Form\PureField;
+use Yiisoft\Form\Field\Factory\PureField;
 
 final class ThemedPureField extends PureField
 {
